move resolveCommonLanguageId
