+++ packages/language-service/index.ts
@@ -4,7 +4,6 @@ export * from './lib/documents';
 export { mergeWorkspaceEdits } from './lib/languageFeatures/rename';
 export * from './lib/types';
 export * as transformer from './lib/transformer';
-export * from './lib/utils/languageId';
 
 import type * as vscode from 'vscode-languageserver-protocol';
