--- packages/language-service/lib/utils/languageId.ts
@@ -1,16 +1,0 @@
-export function resolveCommonLanguageId(fileNameOrUri: string) {
-	const ext = fileNameOrUri.split('.').pop()!;
-	switch (ext) {
-		case 'js': return 'javascript';
-		case 'cjs': return 'javascript';
-		case 'mjs': return 'javascript';
-		case 'ts': return 'typescript';
-		case 'cts': return 'typescript';
-		case 'mts': return 'typescript';
-		case 'jsx': return 'javascriptreact';
-		case 'tsx': return 'typescriptreact';
-		case 'pug': return 'jade';
-		case 'md': return 'markdown';
-	}
-	return ext;
-}
